Added functionality to level and added commands levelup leveldown to main.cc

+++ levelThree.cc
@@ -15,8 +15,8 @@ LevelThree::~LevelThree() {}
 
 AbstractBlock* LevelThree::generateBlock(char type) {
     if (type != ' ') return createCustomBlock(type);
-    const double DEFAULT = 1/9;
-    const double Z_S_PROB = 2/9;
+    const double DEFAULT = static_cast<double>(1)/9;
+    const double Z_S_PROB = static_cast<double>(2)/9;
     char c = generateChar(DEFAULT,DEFAULT,DEFAULT,DEFAULT,
     Z_S_PROB,Z_S_PROB,DEFAULT);
     return createCustomBlock(c);
